requirements.txt
a2wsgi==1.10.10
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
bentoml==1.4.17
cachetools==5.5.2
cattrs==23.1.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-option-group==0.5.7
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.3.0
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fs==2.4.16
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.7.2
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kantoku==0.18.3
kiwisolver==1.4.7
litellm==1.73.6
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
-e git+https://github.com/jjiang10-sv/nyitAge.git@40000cc98eec17d3a2f7fff63496b3248f214c3a#egg=memvid&subdirectory=nyit/RApro/modelTrain/memvid
monotonic==1.6
mpmath==1.3.0
multidict==6.6.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-ml-py==12.575.51
ollama==0.4.9
openai==1.93.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-aiohttp-client==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
packaging==25.0
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
pip-requirements-parser==32.0.1
pipdeptree==2.26.1
platformdirs==4.3.8
pluggy==1.6.0
posthog==3.25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qrcode==8.2
questionary==2.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
safetensors==0.5.3
schema==0.7.7
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
simple-di==0.1.5
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starfish-core==0.1.2
starlette==0.47.1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomli_w==1.2.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
